+++ src/themes/fennec/radar/player/dead-player/dead-player.js
@@ -4,6 +4,12 @@ import { radarConfig } from '/hud/radar/helpers/radar-config.js'
 import { teamColorClass } from '/hud/helpers/team-color-class.js'
 
 export default {
+	data() {
+		return {
+			position: this.player.position,
+		}
+	},
+
 	props: [
 		'player',
 	],
@@ -18,13 +24,13 @@ export default {
 
 		coordinates() {
 			return {
-				x: this.offsetX(this.player.position[0]),
-				y: this.offsetY(this.player.position[1]),
+				x: this.offsetX(this.position[0]),
+				y: this.offsetY(this.position[1]),
 			}
 		},
 
 		level() {
-			return this.getLevel(this.player.position[2])
+			return this.getLevel(this.position[2])
 		},
 	},